+++ src/Simpletools/Db/Elasticsearch/Client.php
@@ -14,7 +14,7 @@ class Client
     protected $___connection;
 
 
-    public function __construct($cluster=null)
+    public function __construct(string|null $cluster=null)
     {
         if($cluster)
             $this->___cluster = $cluster;
@@ -104,7 +104,12 @@ class Client
 			return $this->execute($endpoint, 'DELETE', $data);
 		}
 
-		public function execute($endpoint, $method = 'GET', $data = null,  $contentType ='application/json',$attempt=0, $retryPoints=null)
+		public function execute($endpoint, 
+			$method = 'GET',
+			string|array|null $data = null,
+			$contentType ='application/json',
+			$attempt=0, 
+			string|array|null $retryPoints=null)
 		{
 			$settings = self::$_gSettings[self::$_defaultCluster];